requirements/base.txt
Django==4.2.3
python-decouple==3.8

psycopg2==2.9.5
djangorestframework==3.14.0

django-extensions==3.2.1
django-cors-headers==4.1.0
django-jazzmin==2.6.0
django-jazzmin-admin-rangefilter==1.0.0
django-debug-toolbar==4.2.0
drf-yasg==1.21.7

coreapi==2.3.3

Pillow==10.0.0
requests==2.31.0

boto3==1.26.0
attrs==22.1.0

gunicorn==20.1.0
